Update loot tables (Thanks bugsbugsbugs#0528)
Message from bugsbugsbugs:
I have made some changes to the loot to be more simelar to the original. (I downloaded it from github and used that to make changes.) The following things have been changed:
Cooked porkshop amount is now 1-3 is used to be always 1. In the original I could get up to 3 in one chest.
The amount of apples was always 5, now it is 1-4 I could get up to 4 apples in one chest in the original.
The amount of poison arrows was always 5 and is now 3-5. In the original I sometimes got 3 and sometimes 5 so I made it 3-5.
The amount of arrows in high power is always 8 instead of 5.