Added blocks to the interactables blacklist
Added these new blocks that the player was previously able to interact with:
- Chiseled Bookshelf
- Crafter
- Hopper
- Every currently existing potted plant

Note: Interacting with potted plants creates ghost items. Tested through both localhost and playit.gg for latency. I am unsure about how to fix this since it's a client issue. What's weird is that this bug does not happen to chiseled bookshelves.